restructure processing folder and add pyrosar environment file
