+++ sar_antarctica/nci/processing/pyroSAR/pyrosar_environment.yml
@@ -1,0 +1,185 @@
+name: sar-antarctica-pyrosar
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - aom=3.9.1
+  - argon2-cffi=23.1.0
+  - argon2-cffi-bindings=21.2.0
+  - arrow=1.3.0
+  - babel=2.16.0
+  - bcrypt=4.2.1
+  - blosc=1.21.6
+  - brotli=1.1.0
+  - brotli-bin=1.1.0
+  - brotli-python=1.1.0
+  - bzip2=1.0.8
+  - c-ares=1.34.4
+  - ca-certificates=2024.12.14
+  - certifi=2024.12.14
+  - cffi=1.17.1
+  - charset-normalizer=3.4.1
+  - colour=0.1.5
+  - contourpy=1.3.1
+  - cryptography=44.0.0
+  - cycler=0.12.1
+  - cyrus-sasl=2.1.27
+  - dav1d=1.2.1
+  - dill=0.3.9
+  - fonttools=4.55.3
+  - freetype=2.12.1
+  - freexl=2.0.0
+  - furl=2.1.3
+  - gdal=3.10.0
+  - geoalchemy2=0.13.3
+  - geos=3.13.0
+  - geotiff=1.7.3
+  - giflib=5.2.2
+  - greenlet=3.1.1
+  - h2=4.1.0
+  - hpack=4.0.0
+  - hyperframe=6.0.1
+  - icu=75.1
+  - idna=3.10
+  - importlib-metadata=8.5.0
+  - infinity=1.5
+  - intervals=0.9.2
+  - json-c=0.18
+  - keyutils=1.6.1
+  - kiwisolver=1.4.7
+  - krb5=1.21.3
+  - lcms2=2.16
+  - ld_impl_linux-64=2.43
+  - lerc=4.0.0
+  - libarchive=3.7.7
+  - libavif16=1.1.1
+  - libblas=3.9.0
+  - libbrotlicommon=1.1.0
+  - libbrotlidec=1.1.0
+  - libbrotlienc=1.1.0
+  - libcblas=3.9.0
+  - libcurl=8.11.1
+  - libde265=1.0.15
+  - libdeflate=1.23
+  - libedit=3.1.20240808
+  - libev=4.33
+  - libexpat=2.6.4
+  - libffi=3.4.2
+  - libgcc=14.2.0
+  - libgcc-ng=14.2.0
+  - libgdal-core=3.10.0
+  - libgfortran=14.2.0
+  - libgfortran5=14.2.0
+  - libgomp=14.2.0
+  - libheif=1.19.5
+  - libiconv=1.17
+  - libjpeg-turbo=3.0.0
+  - libkml=1.3.0
+  - liblapack=3.9.0
+  - liblzma=5.6.3
+  - libnghttp2=1.64.0
+  - libnsl=2.0.1
+  - libntlm=1.8
+  - libopenblas=0.3.28
+  - libpng=1.6.45
+  - libpq=17.2
+  - librttopo=1.1.0
+  - libspatialite=5.1.0
+  - libsqlite=3.47.2
+  - libssh2=1.11.1
+  - libstdcxx=14.2.0
+  - libstdcxx-ng=14.2.0
+  - libtiff=4.7.0
+  - libuuid=2.38.1
+  - libwebp-base=1.5.0
+  - libxcb=1.17.0
+  - libxcrypt=4.4.36
+  - libxml2=2.13.5
+  - libxslt=1.1.39
+  - libzlib=1.3.1
+  - lxml=5.3.0
+  - lz4-c=1.10.0
+  - lzo=2.10
+  - matplotlib-base=3.10.0
+  - minizip=4.0.7
+  - multiprocess=0.70.17
+  - munkres=1.1.4
+  - ncurses=6.5
+  - numpy=1.26.4
+  - openjpeg=2.5.3
+  - openldap=2.6.9
+  - openssl=3.4.0
+  - orderedmultidict=1.0.1
+  - packaging=24.2
+  - passlib=1.7.4
+  - pathos=0.3.3
+  - pcre2=10.44
+  - pendulum=3.0.0
+  - phonenumbers=8.13.52
+  - pillow=11.1.0
+  - pip=24.3.1
+  - pox=0.3.5
+  - ppft=1.7.6.9
+  - progressbar2=4.5.0
+  - proj=9.5.1
+  - psycopg2=2.9.9
+  - pthread-stubs=0.4
+  - pycparser=2.22
+  - pyparsing=3.2.1
+  - pyrosar=0.27.0
+  - pysocks=1.7.1
+  - python=3.10.16
+  - python-dateutil=2.9.0.post0
+  - python-utils=3.9.1
+  - python_abi=3.10
+  - pytz=2024.2
+  - pyyaml=6.0.2
+  - qhull=2020.2
+  - rav1e=0.6.6
+  - readline=8.2
+  - requests=2.32.3
+  - setuptools=75.6.0
+  - setuptools-scm=8.1.0
+  - setuptools_scm=8.1.0
+  - shapely=2.0.6
+  - six=1.17.0
+  - snappy=1.2.1
+  - spatialist=0.14.0
+  - sqlalchemy=1.4.54
+  - sqlalchemy-utils=0.41.2
+  - sqlalchemy-utils-arrow=0.41.2
+  - sqlalchemy-utils-babel=0.41.2
+  - sqlalchemy-utils-base=0.41.2
+  - sqlalchemy-utils-color=0.41.2
+  - sqlalchemy-utils-encrypted=0.41.2
+  - sqlalchemy-utils-intervals=0.41.2
+  - sqlalchemy-utils-password=0.41.2
+  - sqlalchemy-utils-pendulum=0.41.2
+  - sqlalchemy-utils-phone=0.41.2
+  - sqlalchemy-utils-timezone=0.41.2
+  - sqlalchemy-utils-url=0.41.2
+  - sqlite=3.47.2
+  - svt-av1=2.3.0
+  - tblib=3.0.0
+  - time-machine=2.16.0
+  - tk=8.6.13
+  - tomli=2.2.1
+  - types-python-dateutil=2.9.0.20241206
+  - typing-extensions=4.12.2
+  - typing_extensions=4.12.2
+  - tzdata=2024b
+  - unicodedata2=15.1.0
+  - uriparser=0.9.8
+  - urllib3=2.3.0
+  - wheel=0.45.1
+  - x265=3.5
+  - xerces-c=3.2.5
+  - xorg-libxau=1.0.12
+  - xorg-libxdmcp=1.1.5
+  - yaml=0.2.5
+  - zipp=3.21.0
+  - zlib=1.3.1
+  - zstandard=0.23.0
+  - zstd=1.5.6
+  - pip:
+    - click==8.1.8
+    - pytest==8.3.4
